+++ Sources/PBXProj/BuildConfigurationList.swift
@@ -9,29 +9,39 @@
 import Foundation
 
 public struct BuildConfigurationList: IsaObject {
-    public let object: [String: Any]
+    public let key: String
+    public let rawObject: [String: Any]
     public let buildConfigurations: [BuildConfiguration]
     public let defaultConfigurationName: String?
     public let defaultConfigurationIsVisible: String
-    public init(_ o: [String: Any], objects: [String: Any]) {
-        self.object = o
+    public init?(key: String, value o: [String: Any], objects: [String: Any]) {
+        guard IsaType(object: o) == .XCConfigurationList else {
+            return nil
+        }
+        self.key = key
+        self.rawObject = o
         self.defaultConfigurationName = o["defaultConfigurationName"] as? String
         let buildConfigurationKeys = o["buildConfigurations"] as! [String]
-        self.buildConfigurations = buildConfigurationKeys.map { key in objects[key] as! [String: Any] }
-            .map(BuildConfiguration.init)
+        self.buildConfigurations = buildConfigurationKeys.map { key in (key, objects) }
+            .flatMap(BuildConfiguration.init)
         self.defaultConfigurationIsVisible = o["defaultConfigurationIsVisible"] as! String
     }
 }
 
 public struct BuildConfiguration: IsaObject {
-    public let object: [String: Any]
+    public let key: String
+    public let rawObject: [String: Any]
     public let name: String
-    public let baseConfigurationReference: String?
+    public let baseConfigurationReference: FileReference?
     public let buildSettings: [String: Any]
-    public init(object o: [String: Any]) {
-        self.object = o
+    public init?(key: String, value o: [String: Any], objects: [String: Any]) {
+        guard IsaType(object: o) == .XCBuildConfiguration else {
+            return nil
+        }
+        self.key = key
+        self.rawObject = o
         self.name = o["name"] as! String
-        self.baseConfigurationReference = o["baseConfigurationReference"] as? String
+        self.baseConfigurationReference = FileReference(key: o["baseConfigurationReference"], objects: objects)
         self.buildSettings = o["buildSettings"] as! [String: Any]
     }
 }