+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## master
+* Introduce `--include-existing` (`--no-include-existing`) option  
+  [Toshihiro Suzuki](https://github.com/toshi0383)
+  [#8](https://github.com/toshi0383/xcconfig-extractor/pull/8)
+
 ## 0.2.0
 
 * Rename `trim-duplicates` to `--no-trim-duplicates`